refactor(cmd): split main modes into packages
