+++ README.md
@@ -53,6 +53,8 @@ LVMSync is organized into modular packages to keep concerns separated:
 - `grpc` – provides the gRPC server and authentication helpers used by the remote daemon.
 - `common` and `internal` – shared helpers and internal utilities such as multi-error handling.
 - `internal/client` – coordinates snapshot preparation and client transfer execution.
+- `cmd/client` – handles snapshot dumping and transport selection.
+- `cmd/apply` – applies streamed data to destination devices.
 - `cmd/lvmsync` – CLI orchestrator with a `signals` subpackage for signal handling and cleanup.
 - `cmd/grpcd` – standalone gRPC daemon exposing LVMSync operations remotely.
 
@@ -62,7 +64,7 @@ This structure allows individual packages to be developed and tested in isolatio
 
 - Snapshot preparation helpers (`ensureVolumeGroups`, `checkDiskSpaceForSnapshot`, `createSnapshotIfNeeded`, `PrepareSnapshot`) and client execution logic are consolidated under `internal/client`.
 - These helpers no longer rely on global variables; configuration and loggers are passed explicitly.
-- `main.go` now orchestrates operations through the `internal/client` package.
+- `main.go` now orchestrates operations through the `cmd/client` and `cmd/apply` packages.
 
 ## Logging
